pyserver/requirements.txt
python-dotenv>=0.21.0
gpiozero>=1.6.2  
picamera>=1.13
paho-mqtt>=1.6.1
firebase-admin==6.0.1 
google-api-core==2.11.0 
google-cloud-storage==2.7.0
SpeechRecognitio>=3.9.0
pyaudio>=0.2.13 